ta: pkcs11: report PKCS11 return code to client
For the PKCS11 TA to provide client a meaningful return code as per
PKCS#11 specification, the GPD TEE command invocation return code
is forced to TEE_SUCCESS and client is expected to read to effective
PKCS11 return code (which can report a success or a failure) from the
PKCS11 TA API.

The first memref argument is used to store this PKCS11 status info
when defined as an input/output memref by the client. In such case,
the output memref is a 32bit buffer that stores the PKCS11 return
code.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>