generic_boot: don't add fdt memory as nsec memory
If the memory is probed from DT, it will include all memory with the
TZRAM. This can lead to the following trace:

  I/TC: Non-secure external DT found
  E/TC:0 0 check_phys_mem_is_outside:330 Non-sec mem (0x10000000:0x40000000) overlaps map (type 2 0x4e000000:0x5d000)
  E/TC:0 0 Panic at core/arch/arm/mm/core_mmu.c:334 <check_phys_mem_is_outside>
  E/TC:0 0 TEE load address @ 0x4e000000
  E/TC:0 0 Call stack:
  E/TC:0 0  0x4e006fd1

Which rightfully complains that the TZRAM lies within the wrongly handed
over overall memory. In the CFG_DT case, all reseved memory regions are
configured by OP-TEE. Additionally this relies on behaviour from U-Boot
to subtract the TZDRAM size from the actual RAM size. This however
contradicts the bindings specification for the memory node which says:

- reg : This property contains all the physical memory ranges of
  your board.

In section 5.d of the booting-without-of.txt file inside the kernel
tree. Since all reserved memory is calculated and reserved by the OP-TEE
generic RAM layout or by platforms themselves, remove the additional
runtime check in the device tree case.

Fixes https://github.com/OP-TEE/optee_os/issues/3567

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>