caam: plat_rng_init without NXP_CAAM_RNG_DRV
In case the CFG_CRYPTO_DRIVER is disabled, but the CAAM is enabled to
retrieve the hardware unique key via the CAAM Master Key Verification
Blob, the CAAM can also be used to retrieve the initial RNG seed for the
OP-TEE PRNG implementation. This allows runtime CAAM support to be
disabled while retaining a hardware generated RNG seed.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>