core: fix handle return in system_open_ta_binary()
Prior to this patch system_open_ta_binary() was not updating the
params[0].value.a out parameter used to carry a handle of the opened
binary. Fix this by assigning the value to the handle to
params[0].value.a.

Note that if only one binary is opened at a time the handle will be 0
and ldelf will still work even without this fix.

Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Jerome Forissier <jerome@forissier.org>
Reported-by: Bastien Simondi <bsimondi@netflix.com>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>