ta: add remote processor firmware ELF parser
Add a ELF parser that will used by the remoteproc generic TA.
It is in charge of parsing the ELF files to extract segments
to load in the remote processor memories.
It also offers an API to retrieve the resource table in the
image. This optional table is defined in a specific section
".resource_table" and is used by the Linux remoteproc driver to
discover services exposed by the remote processor.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>