scripts: add coccinelle support
Coccinelle is a utility tool used to match specific code patterns in
order to modify or statically analyze them.
".cocci"scripts (Semantic Patches) are used to define such patterns and
also to specify what actions have to be performed.
Add script coccicheck, which allows to execute Coccinelle semantic
patches on top of op-tee source code.
It is called using the "coccicheck" target in the op-tee Makefile.
The script invokes semantic patches in 3 fashions:
	- use a single semantic patch: specify and execute a single
	  semantic patch from the local set.
	- use multiple semantic patches: execute all semantic patches
	  from the local set.
	- use a subset of external patches: specify a single or
	  multiple semantic patches, defined outside of the local set.

Reviewed-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Andrea Tagliavini <tagliavini.andrea.1995@gmail.com>