scripts: add GDB script for OP-TEE and QEMU
This script is pre-configured for a standard OP-TEE setup running QEMU
and it also has support for auto-completion when loading different TA's
etc.

There are a set of commands both for loading binaries as well as some
utility commands, for example:
 - connect: attach to the gdb server
 - load_tee: loads tee.elf
 - load_ta <ta_name>: loads a TA (pressing TAB autocompletes)
 - load_host <host_name>: loads host part of example TA (autocompletes)
 - load_tfa <blX>: (TAB autocompletes BL1, BL2 or BL31 etc)
 - load_linux: loads Linux kernel
 - load_uboot: loads U-Boot

Right now it's mainly meant to be used with QEMU, but it should be
possible to extend to also run it with OpenOCD if needed.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>