core: initialize stack canaries as early as possible
In preparation for introducing stack canary checking via compiler
instrumentation, move the initialization of stack canaries as early
as possible, i.e., just after we have set the stack pointer and are
thus able to call a C function.

The debug messages are moved out of the initialization code itself
and are still printed at the same time in the boot sequence, because
the print functions themselves will need the canaries to be fully
initialized when the instrumentation is active.

Signed-off-by: Jerome Forissier <jerome@forissier.org>