core: introduce CFG_CORE_CHECK_STACKS
This commit introduces CFG_CORE_CHECK_STACKS to inspect the stack
canaries using compiler instrumentation (-finstrument-functions). When
enabled, the C compiler will introduce entry and exit hooks in all
functions in the TEE core. On exit thread_check_canaries() is called
and will panic as soon as a stack overflow is detected. Also, the debug
messages showing the addresses of the stack tops and where to put
debugger watchpoints are disabled because they don't seem very useful
anymore: a breakpoint on panic() should work equally well.

Signed-off-by: Jerome Forissier <jerome@forissier.org>